Minor changes in x86

+++ f.hpp
@@ -1,6 +1,6 @@
 #ifndef F_HPP_
 #define F_HPP_
 extern "C" {
-	int f( unsigned char *pBuffer, int W, int H, double A, double B, double C, double D, double S );
+	double f( unsigned char *pBuffer, int W, int H, double A, double B, double C, double D, double S );
 }
 #endif // F_HPP_
